Fix from-scratch init: scale residual output projections by 1/sqrt(2*n_layer)
litgpt's default fresh init omits the 1/sqrt(2*n_layer) scaling for
Llama-style (parallel_residual=False) blocks, so the wide SwiGLU
down-projection amplified the residual stream ~1.5x/layer -> logits ~1000,
loss ~880. Scale attn.proj/mlp.proj. Also cast logits to float32 before
cross_entropy (bf16 CE unsafe).

+++ src/train.py
@@ -110,7 +110,7 @@ def estimate_loss(model, get_val, iters, device):
         x, y = get_val()
         with torch.autocast("cuda", dtype=torch.bfloat16):
             logits = model(x)
-            loss = F.cross_entropy(logits.view(-1, logits.size(-1)), y.view(-1))
+        loss = F.cross_entropy(logits.float().view(-1, logits.size(-1)), y.view(-1))
         losses[k] = loss.item()
     model.train()
     return losses.mean().item()
@@ -210,8 +210,8 @@ def main():
             x, y = get_train()
             with torch.autocast("cuda", dtype=torch.bfloat16):
                 logits = model(x)
-                loss = F.cross_entropy(logits.view(-1, logits.size(-1)), y.view(-1))
-                loss = loss / args.grad_accum
+            loss = F.cross_entropy(logits.float().view(-1, logits.size(-1)), y.view(-1))
+            loss = loss / args.grad_accum
             loss.backward()
             loss_acc += loss.item()
         if args.grad_clip > 0:
